requirements.txt
-i https://pypi.org/simple
pyyaml==5.3.1
multidict==4.7.6; python_version >= '3.5'
ops==0.10.0
